style: isValidResponse 함수 위치 api로 이동
