+++ src/api/checkResponse.ts
@@ -1,0 +1,8 @@
+import { errorResponse, validResponse } from 'api/interface';
+
+export const isValidResponse = (
+  response: validResponse | errorResponse
+): response is validResponse => {
+  const validStatus = [200, 201, 202, 203, 204, 205, 206];
+  return validStatus.includes(response.status);
+};
