+++ src/api/index.ts
@@ -1,2 +1,3 @@
 export * from './refreshableApi';
 export * from './authApi';
+export * from './checkResponse';
